refactor: move updateUser function to user.server.ts and clean up user.service.ts
